requirements.txt
alembic==1.14.0
blinker==1.9.0
boto3==1.35.78
botocore==1.35.78
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.7
MarkupSafe==3.0.2
pdfkit==1.0.0
pillow==11.0.0
pymongo==4.11
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
reportlab==4.2.5
requests==2.32.3
s3transfer==0.10.4
sendgrid==6.11.0
six==1.17.0
SQLAlchemy==2.0.36
starkbank-ecdsa==2.2.0
stripe==11.4.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
